requirements/requirements.txt
Django==4.2
django-extensions==3.2.3
django-filter==24.2
djangorestframework==3.15.1
Markdown==3.6
